% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGmixIRT-package.R
\docType{package}
\name{MGmixIRT-package}
\alias{MGmixIRT}
\alias{MGmixIRT-package}
\title{MGmixIRT: Multigroup Mixture IRT Models for Performance Decline}
\description{
Estimation of multigroup mixture item response theory models for
performance decline in low-stakes assessments: the two-class mixture
model of Bolt, Cohen, and Wollack (2002), the HYBRID model of
Yamamoto (1995), and the multiclass performance decline model of
Jin and Wang (2014), in the general multigroup framework of
List, Robitzsch, Luedtke, Koeller, and Nagy (2017).
}
\references{
Bolt, D. M., Cohen, A. S., & Wollack, J. A. (2002). Item parameter
estimation under conditions of test speededness: Application of a
mixture Rasch model with ordinal constraints. \emph{Journal of
Educational Measurement, 39}(4), 331-348.

Cao, J., & Stokes, S. L. (2008). Bayesian IRT guessing models for
partial guessing behaviors. \emph{Psychometrika, 73}(2), 209-230.

Jin, K.-Y., & Wang, W.-C. (2014). Item response theory models for
performance decline during testing. \emph{Journal of Educational
Measurement, 51}(2), 178-200.

List, M. K., Robitzsch, A., Luedtke, O., Koeller, O., & Nagy, G.
(2017). Performance decline in low-stakes educational assessments:
Different mixture modeling approaches. \emph{Large-scale Assessments
in Education, 5}, 15.

Yamamoto, K. (1995). \emph{Estimating the effects of test length and
test time on parameter estimation using the HYBRID model} (TOEFL
Technical Report TR-10). Educational Testing Service.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cahitgs/MGmixIRT}
  \item Report bugs at \url{https://github.com/cahitgs/MGmixIRT/issues}
}

}
\author{
\strong{Maintainer}: Mehmet Cahit Marangoz \email{cahitgs@gmail.com}

Authors:
\itemize{
  \item Mehmet Cahit Marangoz \email{cahitgs@gmail.com}
}

}
\keyword{internal}
